+++ src/xyz/oihana/schema/constants/traits/products/FeeSpecification.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\products ;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\products\FeeSpecification} class.
+ *
+ * @author  Marc Alcaraz (eKameleon)
+ * @package xyz\oihana\schema\constants\traits\products
+ * @since   1.4.0
+ */
+trait FeeSpecification
+{
+    public const string RATE              = 'rate' ;
+    public const string UNRESOLVED_REASON = 'unresolvedReason' ;
+}